Completed version of HW1 Code
This code has two main methods of computing the inversions.  The first method,
which is not currently used, is the simple brute force approach that you can
implement in order to get an understanding of the (drastic) difference in time
that the algorithm takes to execute.  The second algorithm follows the
suggested recursive sort and merge methodology.  The code will print the sorted
list and the number of inversions in O(nlog(n)) time