Programming solution to HW2
This is the incomplete solution to homework 2.  Currently all of the required
methods are complete, but the solution is incorrect.  It appears that the
arrays are all correctly sorted.  The issue probably exists in the
implementation of the sorter.  The HW question warns that you need to sort the
array exactly how they suggest in the HW.