Answers to homework 3
This is the python implementation of the contraction algorithm.  This
implementation uses a dictionary with each node as its key, and the connections
as the value pair.  Each iteration, a node is selected at random and removed.
One of the nodes previously connected to the removed node is selected and all
connections to the previous node are redirected to that node.  When two nodes
remain, the length of the connections list is examined.  If the length of the
connections list is lower than a global variable containing what was thought to
be the minimum cut, the global variable is updated with the new lowest value.

The algorithm is set to run 4000 times, or n squared.  That is more than enough
iterations to obtain the correct answer.